+++ src/pypi_simple/parse_repo.py
@@ -1,6 +1,7 @@
-from typing import Dict, List, Optional, Tuple, Union
+from typing import Any, Dict, List, Optional, Tuple, Union
 from urllib.parse import urljoin
 from bs4 import BeautifulSoup
+from mailbits import ContentType
 import requests
 from .classes import DistributionPackage, IndexPage, Link, ProjectPage
 from .util import check_repo_version
@@ -108,6 +109,40 @@ def parse_repo_project_page(
     )
 
 
+def parse_repo_project_json(data: Any) -> ProjectPage:
+    """
+    .. versionadded:: 0.10.0
+
+    Parse a project page from an object decoded from an
+    :mimetype:`application/vnd.pypi.simple.v1+json` response (See :pep:`691`).
+    The `~ProjectPage.last_serial` attribute will be set to the value of the
+    ``.meta._last-serial`` field, if any.
+
+    :param data: The decoded body of the JSON response
+    :rtype: ProjectPage
+    :raises TypeError: if ``data`` is not a `dict`
+    :raises UnsupportedRepoVersionError: if the repository version has a
+        greater major component than the supported repository version
+    """
+    if not isinstance(data, dict):
+        raise TypeError("JSON project details response is not a dict")
+    repository_version = data["meta"]["api-version"]
+    check_repo_version(repository_version)
+    try:
+        last_serial = str(data["meta"]["_last-serial"])
+    except KeyError:
+        last_serial = None
+    return ProjectPage(
+        project=data["name"],
+        packages=[
+            DistributionPackage.from_pep691_details(filedata, data["name"])
+            for filedata in data["files"]
+        ],
+        repository_version=repository_version,
+        last_serial=last_serial,
+    )
+
+
 def parse_repo_project_response(project: str, r: requests.Response) -> ProjectPage:
     """
     .. versionadded:: 0.7.0
@@ -120,19 +155,32 @@ def parse_repo_project_response(project: str, r: requests.Response) -> ProjectPa
     :rtype: ProjectPage
     :raises UnsupportedRepoVersionError: if the repository version has a
         greater major component than the supported repository version
+    :raises ValueError: if the response has an unsupported
+        :mailheader:`Content-Type`
     """
-    charset: Optional[str]
-    if "charset" in r.headers.get("content-type", "").lower():
-        charset = r.encoding
+    ct = ContentType.parse(r.headers.get("content-type", "text/html"))
+    if ct.content_type == "application/vnd.pypi.simple.v1+json":
+        page = parse_repo_project_json(r.json())
+    elif (
+        ct.content_type == "application/vnd.pypi.simple.v1+html"
+        or ct.content_type == "text/html"
+    ):
+        charset: Optional[str]
+        if "charset" in ct.params:
+            charset = r.encoding
+        else:
+            charset = None
+        page = parse_repo_project_page(
+            project=project,
+            html=r.content,
+            base_url=r.url,
+            from_encoding=charset,
+        )
     else:
-        charset = None
-    page = parse_repo_project_page(
-        project=project,
-        html=r.content,
-        base_url=r.url,
-        from_encoding=charset,
-    )
-    return page._replace(last_serial=r.headers.get("X-PyPI-Last-Serial"))
+        raise ValueError(f"Response has unsupported Content-Type {str(ct)!r}")
+    if page.last_serial is None:
+        page = page._replace(last_serial=r.headers.get("X-PyPI-Last-Serial"))
+    return page
 
 
 def parse_repo_index_page(
@@ -162,6 +210,36 @@ def parse_repo_index_page(
     )
 
 
+def parse_repo_index_json(data: Any) -> IndexPage:
+    """
+    .. versionadded:: 0.10.0
+
+    Parse an index/root page from an object decoded from an
+    :mimetype:`application/vnd.pypi.simple.v1+json` response (See :pep:`691`).
+    The `~IndexPage.last_serial` attribute will be set to the value of the
+    ``.meta._last-serial`` field, if any.
+
+    :param data: The decoded body of the JSON response
+    :rtype: IndexPage
+    :raises UnsupportedRepoVersionError: if the repository version has a
+        greater major component than the supported repository version
+    :raises TypeError: if ``data`` is not a `dict`
+    """
+    if not isinstance(data, dict):
+        raise TypeError("JSON project list response is not a dict")
+    repository_version = data["meta"]["api-version"]
+    check_repo_version(repository_version)
+    try:
+        last_serial = str(data["meta"]["_last-serial"])
+    except KeyError:
+        last_serial = None
+    return IndexPage(
+        projects=[p["name"] for p in data["projects"]],
+        repository_version=repository_version,
+        last_serial=last_serial,
+    )
+
+
 def parse_repo_index_response(r: requests.Response) -> IndexPage:
     """
     .. versionadded:: 0.7.0
@@ -173,11 +251,24 @@ def parse_repo_index_response(r: requests.Response) -> IndexPage:
     :rtype: IndexPage
     :raises UnsupportedRepoVersionError: if the repository version has a
         greater major component than the supported repository version
+    :raises ValueError: if the response has an unsupported
+        :mailheader:`Content-Type`
     """
-    charset: Optional[str]
-    if "charset" in r.headers.get("content-type", "").lower():
-        charset = r.encoding
+    ct = ContentType.parse(r.headers.get("content-type", "text/html"))
+    if ct.content_type == "application/vnd.pypi.simple.v1+json":
+        page = parse_repo_index_json(r.json())
+    elif (
+        ct.content_type == "application/vnd.pypi.simple.v1+html"
+        or ct.content_type == "text/html"
+    ):
+        charset: Optional[str]
+        if "charset" in ct.params:
+            charset = r.encoding
+        else:
+            charset = None
+        page = parse_repo_index_page(html=r.content, from_encoding=charset)
     else:
-        charset = None
-    page = parse_repo_index_page(html=r.content, from_encoding=charset)
-    return page._replace(last_serial=r.headers.get("X-PyPI-Last-Serial"))
+        raise ValueError(f"Response has unsupported Content-Type {str(ct)!r}")
+    if page.last_serial is None:
+        page = page._replace(last_serial=r.headers.get("X-PyPI-Last-Serial"))
+    return page