+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 v0.10.0 (in development)
 ------------------------
 - Support Python 3.10
+- Support PEP 691
+    - Requests now include "Accept" headers that prefer the new JSON format
+      over the old HTML format
+    - `parse_repo_project_response()` and `parse_repo_index_response()` now
+      support both the JSON and HTML formats
+    - Add `parse_repo_index_json()` and `parse_repo_project_json()` functions
+    - Gave `DistributionPackage` a `digests` attribute for storing the contents
+      of the `"hashes"` field in JSON project details
+    - Gave `DistributionPackage` a `from_pep691_details()` classmethod
 
 v0.9.0 (2021-08-26)
 -------------------